# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/messages/send": {
            "post": {
                "tags": [
                    "Send Messages"
                ],
                "description": "Sending Messages in the system",
                "operationId": "816a9a107d1c7ab2f9b04af172c561b3",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {token}",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "receiver_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "message"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sending Messages in the system",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sender_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "receiver_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "body": {
                                            "type": "string",
                                            "example": "Hello John Doe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messages/{receiverId}": {
            "get": {
                "tags": [
                    "Receive Message"
                ],
                "description": "Receiving Messages",
                "operationId": "bd8775dbf34a3a0799f93f7a784c3040",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {token}",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receiving Messages in the system",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "hello 1 , hello 2"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messages/{id}/read": {
            "patch": {
                "tags": [
                    "Mark Messages as read"
                ],
                "description": "Mark messages as read",
                "operationId": "42e15a6796c70cd699eb810186a73508",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {token}",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updating messages as read messages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mark Message as read"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}